refactor: move classes for better typing
